Admits caveat of SIGTERM handling in InfraValidator.
Context:
In the explicit graceful shutdown beahvior of InfraValidator, we registered a SIGTERM handler that raises exception so that cleanup code in "finally" block can cover the signal as well. However it turns out that threads other than MainThread is NOT allowed to register the signal handler. If the executor runs on worker thread, signal handler registration will fail.

Decision:
InfraValidator will not handle SIGTERM if the executor thread is not a MainThread.

Reason:
For Beam DirectRunner, as Beam does not support graceful shutdown on each task (that runs each component in BeamDagRunner), there's nothing InfraVaildator can do.
Even without graceful shutdown behavior, external resources (such as model server Pod) can be cleaned up with platform-native mechanism (activeDeadlineSeconds), so this explicit grace period behavior is just an enhancment feature.
PiperOrigin-RevId: 306203027